refactored do_discover into discover.py file
