+++ tap_trello/discover.py
@@ -1,0 +1,47 @@
+import os
+import json
+
+from singer import metadata
+from singer.catalog import Catalog
+from .streams import STREAM_OBJECTS
+
+
+def _get_abs_path(path):
+    return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
+
+
+# Load schemas from schemas folder
+def _load_schemas():
+    schemas = {}
+
+    for filename in os.listdir(_get_abs_path("schemas")):
+        path = _get_abs_path("schemas") + "/" + filename
+        file_raw = filename.replace(".json", "")
+        with open(path) as file:
+            schemas[file_raw] = json.load(file)
+
+    return schemas
+
+
+def do_discover():
+    raw_schemas = _load_schemas()
+    catalog_entries = []
+
+    for stream_name, schema in raw_schemas.items():
+        # create and add catalog entry
+        stream = STREAM_OBJECTS[stream_name]
+        catalog_entry = {
+            "stream": stream_name,
+            "tap_stream_id": stream_name,
+            "schema": schema,
+            "metadata": metadata.get_standard_metadata(
+                schema=schema,
+                key_properties=stream.key_properties,
+                valid_replication_keys=stream.replication_keys,
+                replication_method=stream.replication_method,
+            ),
+            "key_properties": stream.key_properties,
+        }
+        catalog_entries.append(catalog_entry)
+
+    return Catalog.from_dict({"streams": catalog_entries})
